backend/requirements.txt
# Invoice Automation MVP - Backend
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
cryptography==41.0.7
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
gunicorn==21.2.0
Werkzeug==3.0.1
pytest==7.4.3
